Augment an IEA data frame

This function prepares an IEA data frame created by iea_df() for use in R.

augment_iea_df(.iea_df, country = "Country",
  ledger_side = "Ledger.side",
  flow_aggregation_point = "Flow.aggregation.point", flow = "Flow",
  energy_type = "Energy.type", energy_type_val = "E",
  units = "Units", units_val = "ktoe", supply = "Supply",
  consumption = "Consumption", tpes = "Total primary energy supply",
  tpes_flows = c("Production", "Imports", "Exports",
  "International marine bunkers", "International aviation bunkers",
  "Stock changes"), tfc_compare = "TFC compare",
  tfc_compare_flows = c("Total primary energy supply", "Transfers",
  "Statistical differences", "Transformation processes",
  "Energy industry own use", "Losses"), losses = "Losses",
  transformation_processes = "Transformation processes",
  tp_flows_suffix = "(transf.)",
  nstp_flows_suffix = "(transformation)",
  eiou = "Energy industry own use", eiou_flows_suffix = "(energy)",
  tfc = "Total final consumption", tfc_flows = c("Industry",
  "Transport", "Other", "Non-energy use"), industry = "Industry",
  industry_flows = c("Iron and steel", "Chemical and petrochemical",
  "Non-ferrous metals", "Non-metallic minerals", "Transport equipment",
  "Machinery", "Mining and quarrying", "Food and tobacco",
  "Paper, pulp and print", "Wood and wood products", "Construction",
  "Textile and leather", "Non-specified (industry)"),
  transport = "Transport",
  transport_flows = c("World aviation bunkers", "Domestic aviation",
  "Road", "Rail", "Pipeline transport", "World marine bunkers",
  "Domestic navigation", "Non-specified (transport)"), other = "Other",
  other_flows = c("Residential", "Commercial and public services",
  "Agriculture/forestry", "Fishing", "Non-specified (other)"),
  non_energy = "Non-energy use", non_energy_prefix = "Non-energy use",
  electricity_output = "Electricity output (GWh)",
  electricity_output_flows_prefix = "Electricity output (GWh)-",
  heat_output = "Heat output",
  heat_output_flows_prefix = "Heat output-", .rownum = ".rownum")

Arguments

.iea_df
a data frame produced by the iea_df() function

country
the name of the country column in .iea_df. Default is "Country".

ledger_side
the name of the ledger side column to be added to .iea_df. Default is "Ledger.side".

flow_aggregation_point
the name of the flow aggregation point column to be added to .iea_df. Default is "Flow.aggregation.point".

flow
the name of the flow column in .iea_df. Default is "Flow".

energy_type
the name of the energy type column to be added to .iea_df. Default is "Energy.type.

energy_type_val
the value to put in the energy_type column. Default is "E".

units
the name of the units column to be added to .iea_df. Default is "Units".

units_val
the value to put in the units column. Default is "ktoe" for kilotons of oil equivalent.

supply
the string that identifies supply Ledger.side. Default is "Supply".

consumption
the string that identifies consumption Ledger.side. Default is "Consumption".

tpes
the string that identifies total primary energy supply Flow.aggregation.point. Default is "Total primary energy supply.

tpes_flows
a vector of strings that give flows that are aggregated to Total primary energy supply.

tfc_compare
a string that identifies the TFC compare flow aggregation point. Default is TFC compare.

tfc_compare_flows
a vector of strings that give Flows that are aggregated to TFC compare.

losses
the string that indicates losses in the Flow column. Default is "Losses".

transformation_processes
the string that indicates transformation processes in the Flow column. Default is "Transformation processes".

tp_flows_suffix
the suffix for transformation processes in the Flow column. Default is "(transf.)".

nstp_flows_suffix
the suffix for non-specified transformation processes in the Flow column. Default is "(transformation)".

eiou
the string that identifies energy industry own use in the Flow column. Default is "Energy industry own use".

eiou_flows_suffix
the suffix for energy industry own use in the Flow column. Default is "(energy)".

tfc
the string that identifies total final consumption in the Flow column. Default is "Total final consumption".

tfc_flows
a vector of strings that give total final consumption in the Flow column.

industry
a string that names the industry Flow.aggregation.point. Default is "Industry".

industry_flows
a vector of strings representing Flows to be aggregated in the Industry Flow.aggregation.point.

transport
a string that names the transport Flow.aggregation.point. Default is "Transport".

transport_flows
a vector of strings representing Flows to be aggregated in the Transport Flow.aggregation.point.

other
a string that names the other Flow.aggregation.point. Default is "Other".

other_flows
a vector of strings representing Flows to be aggregated in the Other Flow.aggregation.point.

non_energy
a string that names the non-energy Flow.aggregation.point. Default is "Non-energy use".

non_energy_prefix
a string prefix for Flows to be aggregated in the Non-energy use Flow.aggregation.point.

electricity_output
a string that names the electricity output Flow. Default is "Electricity output (GWh)".

electricity_output_flows_prefix
a string prefix for Flows to be aggregated in electricity output. Default is "Electricity output (GWh)-".

heat_output
a string that names the heat output Flow. Default is "Heat output".

heat_output_flows_prefix
a string prefix for Flows to be aggregated in heat output. Default is "Heat output-".

.rownum
the name of a column created (and destroyed) internally by this function. The .rownum column temporarily holds row numbers for internal calculations. The .rownum column is deleted before returning.

Value

.iea_df with additional columns named ledger_side, flow_aggregation_point, energy_type, and units.

Details

This function solves several problems. The first problem is that metadata in the COUNTRY, FLOW, and PRODUCT columns of an IEA data table are not unique. To solve this problem, two additional columns are added: Ledger.side and Flow.aggregation.point. Ledger.side can be one of "Supply" or "Consumption", corresponding to the top or bottom of the IEA's tables, respectively. Flow.aggregation.point indicates the next level of aggregation for these data. Flow.aggregation.point can be one of "Total primary energy supply", "Transformation processes", "Energy industry own use", or "TFC compare" on the Supply side of the ledger. On the Consumption side of the ledger, Flow.aggregation.point can be one of "Industry", "Transport", "Other", or "Non-energy use".

The second problem this function solves is that energy type and units are not specified in IEA data. An Energy.type column is added with the value of energy_type_val. (Default is E, for energy, as opposed to X, which would be exergy.) A Units column is added with the value of units_val. (Default is ktoe, although any string can be specified in units_val.)

Note that this function decides where to divide Supply from Consumption. To do so, it first looks for rows in which Flow is "Losses". The last "Losses" row is the last row of the Supply side of the ledger. If "Losses" rows are not found, the function looks for rows in which Flow is "Total final consumption". The first "Total final consumption" row is the first row of the Consumption side of the ledger. If neither "Losses" nor "Total final consumption" Flows are present, an error is generated.

Examples

iea_df(text = paste0(",,TIME,1960,1961\n",
                       "COUNTRY,FLOW,PRODUCT\n",
                       "World,Production,Hard coal (if no detail),42,43\n",
                       "World,Losses,Hard coal (if no detail),1,2")) %>%
  rename_iea_df_cols() %>%
  augment_iea_df()

#> # A tibble: 2 x 9
#>   Country Ledger.side Flow.aggregatio… Energy.type Units Flow  Product `1960`
#>   <chr>   <chr>       <chr>            <chr>       <chr> <chr> <chr>    <dbl>
#> 1 World   Supply      Total primary e… E           ktoe  Prod… Hard c…     42
#> 2 World   Supply      TFC compare      E           ktoe  Loss… Hard c…      1
#> # … with 1 more variable: `1961` <dbl>